small_migration/ai-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.10
pandas>=2.1.4
openai-agents>=0.0.7
httpx>=0.27.0
